+++ app/src/main/java/com/vci/vectorcamapp/imaging/domain/camera/CameraMetadataListener.kt
@@ -1,0 +1,8 @@
+package com.vci.vectorcamapp.imaging.domain.camera
+
+import com.vci.vectorcamapp.imaging.domain.model.CameraMetadata
+
+interface CameraMetadataListener {
+    val latestMetadata: CameraMetadata?
+    fun reset()
+}
