+++ app/src/main/java/com/vci/vectorcamapp/imaging/presentation/ImagingScreen.kt
@@ -1,9 +1,5 @@
 package com.vci.vectorcamapp.imaging.presentation
 
-import android.hardware.camera2.CameraCaptureSession
-import android.hardware.camera2.CaptureRequest
-import android.hardware.camera2.CaptureResult
-import android.hardware.camera2.TotalCaptureResult
 import android.view.Surface
 import androidx.camera.camera2.interop.Camera2Interop
 import androidx.camera.core.Camera
@@ -15,7 +11,6 @@ import androidx.camera.core.SurfaceRequest
 import androidx.camera.core.resolutionselector.AspectRatioStrategy
 import androidx.camera.core.resolutionselector.ResolutionSelector
 import androidx.camera.lifecycle.ProcessCameraProvider
-import java.util.concurrent.atomic.AtomicReference
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -78,6 +73,7 @@ import com.vci.vectorcamapp.core.presentation.components.empty.EmptySpace
 import com.vci.vectorcamapp.core.presentation.components.form.TextEntryField
 import com.vci.vectorcamapp.core.presentation.components.form.ToggleField
 import com.vci.vectorcamapp.core.presentation.components.tile.InfoTile
+import com.vci.vectorcamapp.imaging.data.camera.CameraMetadataListenerImplementation
 import com.vci.vectorcamapp.imaging.presentation.components.camera.LiveCameraPreview
 import com.vci.vectorcamapp.imaging.presentation.components.icon.AnimatedArrowIcon
 import com.vci.vectorcamapp.imaging.presentation.components.specimen.CapturedSpecimenTile
@@ -101,7 +97,7 @@ fun ImagingScreen(
     var surfaceRequest by remember { mutableStateOf<SurfaceRequest?>(null) }
     var imageCaptureUseCase by remember { mutableStateOf<ImageCapture?>(null) }
     var camera by remember { mutableStateOf<Camera?>(null) }
-    val latestCaptureMetadata = remember { AtomicReference<CameraMetadata?>(null) }
+    val metadataListener = remember { CameraMetadataListenerImplementation() }
 
     val isReviewing by rememberUpdatedState(newValue = state.currentImageBytes != null)
     val analyzer = remember {
@@ -147,41 +143,7 @@ fun ImagingScreen(
             )
 
         Camera2Interop.Extender(imageCaptureBuilder)
-            .setSessionCaptureCallback(object : CameraCaptureSession.CaptureCallback() {
-                override fun onCaptureCompleted(
-                    session: CameraCaptureSession,
-                    request: CaptureRequest,
-                    result: TotalCaptureResult
-                ) {
-                    val rggb = result[CaptureResult.COLOR_CORRECTION_GAINS]
-                    latestCaptureMetadata.set(
-                        CameraMetadata(
-                            focusDistance = result[CaptureResult.LENS_FOCUS_DISTANCE],
-                            aperture = result[CaptureResult.LENS_APERTURE],
-                            exposureTimeNs = result[CaptureResult.SENSOR_EXPOSURE_TIME],
-                            iso = result[CaptureResult.SENSOR_SENSITIVITY],
-                            colorCorrectionGains = rggb?.let {
-                                ColorCorrectionGains(
-                                    red = it.red,
-                                    greenEven = it.greenEven,
-                                    greenOdd = it.greenOdd,
-                                    blue = it.blue
-                                )
-                            },
-                            awbMode = result[CaptureResult.CONTROL_AWB_MODE],
-                            afRegions = result[CaptureResult.CONTROL_AF_REGIONS]?.map { region ->
-                                AfRegion(
-                                    x = region.x,
-                                    y = region.y,
-                                    width = region.width,
-                                    height = region.height,
-                                    weight = region.meteringWeight
-                                )
-                            } ?: emptyList()
-                        )
-                    )
-                }
-            })
+            .setSessionCaptureCallback(metadataListener)
 
         val imageCapture = imageCaptureBuilder.build()
 
@@ -774,7 +736,7 @@ fun ImagingScreen(
                                         label = "Capture",
                                         onClick = {
                                             imageCaptureUseCase?.let {
-                                                onAction(ImagingAction.CaptureImage(it, latestCaptureMetadata.get()))
+                                                onAction(ImagingAction.CaptureImage(it, metadataListener.latestMetadata))
                                             }
                                         },
                                         iconPainter = painterResource(id = R.drawable.ic_camera),
